Adds keystore and cleans up manifest
Cleans up manifest, removing entries for activities that no longer exist
and eliminates additional "apps" in android that shouldn't be visible
outside of Minder (they're regular activities)

Additionally creates the keystore that can be used to sign the application
for release. The password is the first word of the course name - ________
computing.